Rework decodeSampleDS For the decodeSampleDS function, we are now using src_short_to_float_array to do the main conversion and calling std::reverse to reverse if necessary.
Since the DrumSynth::GetDSFileSamples function calls
new on the out parameter "wave", we are using std::unique_ptr
to manage the lifetime without explicitly calling delete.

We now throw an error if the DrumSynth file could not be decoded,
which seems like the most reasonable to do in this case.
loadFromAudioFile currently bails for audio files that failed to load,
but this may be changed depending on if the bailing behavior is wanted.