Remove original data and undo changes to recording files The only change made to the recording files was to set the sampleRate within the constructor.
We were also attempting to always resample the buffer to engine,
which may help performance somewhat when being played back,
but it's more useful to leave the buffer in the intended sample
rate to avoid surprises. Sample rate conversion on playback is handled
by Sample, so this is not an issue.