Remove audio FIFO thread
Was an extra layer of indirection that did seem to help latency much, but made
the design more complex.

The frames per period will thus always be
DEFAULT_BUFFER_SIZE as well as not being
influenced by a FIFO size. This means that the frames per period wont be read in from the config, but this fixed size was already being enforced for most of the codebase anyways.

In the future, frames per period should
be read in from the config and then
a device should be made from that if
possible. If not, then we can default
to using DEFAULT_BUFFER_SIZE.